ListConqueredPlacesTest CI 성공

+++ app-server/subprojects/bounded_context/place_search/infra/src/integrationTest/kotlin/club/staircrusher/place_search/infra/adapter/in/controller/ListConqueredPlacesTest.kt
@@ -28,7 +28,7 @@ class ListConqueredPlacesTest : PlaceSearchITBase() {
             .sccRequest("/listConqueredPlaces", requestBody = null, user = user)
             .apply {
                 val result = getResult(object : TypeReference<ListConqueredPlacesResponseDto>() {})
-                assertEquals(registeredCount, result.totalNumberOfItems)
+                assertEquals(registeredCount.toLong(), result.totalNumberOfItems)
                 result.items.forEach { item ->
                     val place = places.find { it.id == item.place.id }!!
                     assertEquals(place.building.id, item.building.id)
